pyproject.toml
[project]
name = "meetup_bot"
version = "1.1.2"
description = "Use Meetup Pro API to send Slack messages before events occur."
authors = [
    { name = "pythoninthegrass", email = "4097471+pythoninthegrass@users.noreply.github.com" }
]
readme = "README.md"
license = "Unlicense"

requires-python = ">=3.11,<3.12"

dependencies = [
    "arrow>=1.3.0,<2",
    "bcrypt==4.0.1",
    "colorama>=0.4.5,<0.5",
    "exceptiongroup>=1.2.2,<2",
    "fastapi>=0.115.6",
    "gunicorn>=23.0.0,<24",
    "hishel>=1.1.9",
    "httpx>=0.28.1",
    "icecream>=2.1.3,<3",
    "Jinja2>=3.1.5,<4",
    "jose>=1.0.0,<2",
    "numpy>=2.2.1,<3",
    "pandas>=2.2.3,<3",
    "pony>=0.7.16,<0.8",
    "PyJWT[crypto]>=2.10.1,<3",
    "python-decouple~=3.8",
    "python-jose[cryptography]>=3.3.0,<4",
    "python-multipart>=0.0.9,<0.0.10",
    "slack-sdk>=3.34.0,<4",
    "uvicorn>=0.29.0,<0.30",
    "wheel>=0.43.0,<0.44",
]

[project.optional-dependencies]
dev = [
    "argcomplete<4.0.0,>=3.5.0",
    "deptry<1.0.0,>=0.23.0",
    "icecream<3.0.0,>=2.1.3",
    "ipython<9.0.0,>=8.27.0",
    "mypy<2.0.0,>=1.14.1",
    "pyclean<4.0.0,>=3.0.0",
    "pytest-asyncio<1.0.0,>=0.25.2",
    "pytest-cov<7.0.0,>=6.0.0",
    "pytest<9.0.0,>=8.3.4",
    "rich<14.0.0,>=13.8.1",
    "ruff>=0.9.5",
]
test = [
    "coverage<8.0.0,>=7.6.1",
    "hypothesis[cli]<7.0.0,>=6.112.1",
    "pytest<9.0.0,>=8.3.3",
    "pytest-asyncio<1.0.0,>=0.24.0",
    "pytest-cov>=6.1.1",
    "pytest-datafiles<4.0.0,>=3.0.0",
    "pytest-xdist<4.0.0,>=3.6.1",
]

[tool.pytest.ini_options]
pythonpath = ["app"]
asyncio_mode = "auto"
markers = [
    "unit: unit tests (no external dependencies)",
    "integration: integration tests (component interactions, TestClient)",
    "e2e: end-to-end tests (requires running server)",
    "property: property-based tests (Hypothesis)",
]

[tool.deptry]
# DEP003: transitive deps
ignore = [
    "DEP003"
]

[tool.deptry.per_rule_ignores]
# DEP002: not used in codebase (excluding dev deps)
DEP002 = [
    "deptry",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "uvicorn"
]


